% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSR_monitor}
\alias{LSR_monitor}
\title{Lilaeopsis schaffneriana var. recurva monitoring data, 2007, 2010-2016}
\format{A data frame with 152 rows and 5 variables
\describe{
  \item{\code{Year}}{Year monitoring was conducted}
  \item{\code{Site}}{Site at which monitoring was conducted}
  \item{\code{Patch}}{A patch identifier to allow syncing the estimates
    made by different observers}
  \item{\code{Observer}}{A unique identifier within year for the observer}
  \item{\code{Area}}{The estimated area of the patch, given the observer's
    sketch and their measurements}
}}
\source{
San Bernardino and Leslie Canyon NWRs, US Fish and Wildlife Service
}
\usage{
LSR_monitor
}
\description{
MORE DESCRIPTION
}
\keyword{datasets}

