% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_vignette}
\alias{run_vignette}
\title{Run the vignette for the LSR monitoring analysis}
\usage{
run_vignette()
}
\value{
An HTML version of the LSR analysis vignette
}
\description{
Run the vignette for the LSR monitoring analysis
}
\examples{
run_vignette()
}

