% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_plot}
\alias{summary_plot}
\title{Summary plots of some sort}
\usage{
summary_plot(x)
}
\arguments{
\item{paramName}{explanation}
}
\value{
what is returned
}
\description{
Summary plots of some sort
}
\examples{
one or more lines to demo the function
}
\seealso{
if any see alsos
}

